[LEGACY] Replaced old inventory management modules
CoroutineStealer -> ChestStealer
CoroutineCleaner -> InventoryCleaner
CoroutineArmorer -> AutoArmor
CoroutineArmorComparator -> ArmorComparator

Added missing headers.
