+++ src/main/java/net/ccbluex/liquidbounce/api/ClientSettings.kt
@@ -1,3 +1,8 @@
+/*
+ * LiquidBounce Hacked Client
+ * A free open source mixin-based injection hacked client for Minecraft using Minecraft Forge.
+ * https://github.com/CCBlueX/LiquidBounce/
+ */
 package net.ccbluex.liquidbounce.api
 
 import net.ccbluex.liquidbounce.utils.ClientUtils.LOGGER
