+++ src/main/java/net/ccbluex/liquidbounce/features/special/AutoReconnect.kt
@@ -3,7 +3,6 @@
  * A free open source mixin-based injection hacked client for Minecraft using Minecraft Forge.
  * https://github.com/CCBlueX/LiquidBounce/
  */
-
 package net.ccbluex.liquidbounce.features.special
 
 object AutoReconnect {
