+++ src/main/java/net/ccbluex/liquidbounce/lang/Language.kt
@@ -1,3 +1,8 @@
+/*
+ * LiquidBounce Hacked Client
+ * A free open source mixin-based injection hacked client for Minecraft using Minecraft Forge.
+ * https://github.com/CCBlueX/LiquidBounce/
+ */
 package net.ccbluex.liquidbounce.lang
 
 import net.ccbluex.liquidbounce.file.FileManager.PRETTY_GSON
