+++ src/main/java/net/ccbluex/liquidbounce/utils/inventory/ArmorComparator.kt
@@ -1,3 +1,8 @@
+/*
+ * LiquidBounce Hacked Client
+ * A free open source mixin-based injection hacked client for Minecraft using Minecraft Forge.
+ * https://github.com/CCBlueX/LiquidBounce/
+ */
 package net.ccbluex.liquidbounce.utils.inventory
 
 import net.ccbluex.liquidbounce.utils.MinecraftInstance
@@ -7,7 +12,7 @@ import net.minecraft.item.ItemArmor
 import net.minecraft.item.ItemStack
 import kotlin.math.ceil
 
-object CoroutineArmorComparator: MinecraftInstance() {
+object ArmorComparator: MinecraftInstance() {
 	fun getBestArmorSet(stacks: List<ItemStack?>, entityStacksMap: Map<ItemStack, EntityItem>? = null): ArmorSet? {
 		val thePlayer = mc.thePlayer ?: return null
