--- src/main/java/net/ccbluex/liquidbounce/utils/inventory/ArmorPiece.java
@@ -1,32 +1,0 @@
-/*
- * LiquidBounce Hacked Client
- * A free open source mixin-based injection hacked client for Minecraft using Minecraft Forge.
- * https://github.com/CCBlueX/LiquidBounce/
- */
-package net.ccbluex.liquidbounce.utils.inventory;
-
-
-import net.minecraft.item.ItemArmor;
-import net.minecraft.item.ItemStack;
-
-public class ArmorPiece {
-    private final ItemStack itemStack;
-    private final int slot;
-
-    public ArmorPiece(ItemStack itemStack, int slot) {
-        this.itemStack = itemStack;
-        this.slot = slot;
-    }
-
-    public int getArmorType() {
-        return ((ItemArmor) itemStack.getItem()).armorType;
-    }
-
-    public int getSlot() {
-        return slot;
-    }
-
-    public ItemStack getItemStack() {
-        return itemStack;
-    }
-}
