+++ src/main/java/net/ccbluex/liquidbounce/utils/timing/DelayTimer.kt
@@ -1,3 +1,8 @@
+/*
+ * LiquidBounce Hacked Client
+ * A free open source mixin-based injection hacked client for Minecraft using Minecraft Forge.
+ * https://github.com/CCBlueX/LiquidBounce/
+ */
 package net.ccbluex.liquidbounce.utils.timing
 
 import net.ccbluex.liquidbounce.utils.inventory.InventoryUtils.CLICK_TIMER
